requirements.txt
absl-py==0.2.2
addict==2.1.3
arrow==0.12.1
astor==0.6.2
atomicwrites==1.1.5
attrs==18.1.0
bleach==1.5.0
boto3==1.7.55
botocore==1.10.55
certifi==2018.4.16
chardet==3.0.4
cndevspaces==0.2.19
colorama==0.3.9
cycler==0.10.0
cymem==1.31.2
cytoolz==0.8.2
dill==0.2.8.1
docker==3.4.0
docker-pycreds==0.3.0
docutils==0.14
en-core-web-sm==2.0.0
ftfy==5.4.1
gast==0.2.0
gitdb2==2.0.3
GitPython==2.1.10
graypy==0.3
grpcio==1.12.1
html5lib==0.9999999
idna==2.7
jmespath==0.9.3
joblib==0.11
kiwisolver==1.0.1
Markdown==2.6.11
marshmallow==2.15.1
matplotlib==2.2.2
more-itertools==4.2.0
msgpack-numpy==0.4.1
msgpack-python==0.5.6
murmurhash==0.28.0
numpy==1.14.5
pandas==0.23.1
pathlib==1.0.1
plac==0.9.6
pluggy==0.6.0
preshed==1.0.0
protobuf==3.6.0
psutil==5.4.6
py==1.5.4
pyparsing==2.2.0
pytest==3.6.3
python-dateutil==2.7.3
pytz==2018.4
PyYAML==3.13
regex==2017.4.5
requests==2.19.1
retrying==1.3.3
s3transfer==0.1.13
scikit-learn==0.19.1
scipy==1.1.0
simplejson==3.16.0
six==1.10.0
sklearn==0.0
smmap2==2.0.3
spacy==2.0.11
syncthing==2.1.2
syncthingmanager==0.1.0
tabulate==0.8.2
tensorboard==1.8.0
tensorflow-gpu==1.8.0
termcolor==1.1.0
thinc==6.10.2
toolz==0.9.0
torch==0.4.0
tqdm==4.23.4
ujson==1.35
urllib3==1.23
wcwidth==0.1.7
websocket-client==0.48.0
Werkzeug==0.14.1
wrapt==1.10.11
